refactor: relocate Callout component
