+++ src/mdx-components.tsx
@@ -1,9 +1,9 @@
 import defaultMdxComponents from 'fumadocs-ui/mdx';
-import {Step, Steps} from 'fumadocs-ui/components/steps';
-import {Accordion, Accordions} from 'fumadocs-ui/components/accordion';
-import type {MDXComponents} from 'mdx/types';
+import { Step, Steps } from 'fumadocs-ui/components/steps';
+import { Accordion, Accordions } from 'fumadocs-ui/components/accordion';
+import type { MDXComponents } from 'mdx/types';
 import * as TabsComponents from 'fumadocs-ui/components/tabs';
-import Callout from '@/components/Pulse/Callout/Callout'
+import Callout from '@/components/Pulse/Callout'
 import { LinkIcon } from 'lucide-react';
 import Metric from './components/Metric/Metric'
 import ZoomImage from '@/components/Pulse/ZoomImage'
@@ -18,7 +18,7 @@ export function getMDXComponents(components?: MDXComponents): MDXComponents {
     Accordion,
     Accordions,
     Callout,
-    Metric: ({...props}) => <Metric {...props} className='my-6'/>,
+    Metric: ({ ...props }) => <Metric {...props} className='my-6' />,
     h1: ({ id, children, ...props }) => (
       <h1
         id={id}
@@ -39,7 +39,7 @@ export function getMDXComponents(components?: MDXComponents): MDXComponents {
       </h1>
     ),
     img: (props) => {
-      const {src, alt, ...rest} = props;
+      const { src, alt, ...rest } = props;
       const imageSrc = typeof src === 'object' && src !== null ? (src as { src: string }).src : src;
       if (!imageSrc) return null;
